+++ crates/bevy_cef_core/src/browser_process/browsers.rs
@@ -36,9 +36,13 @@ mod keyboard;
 
 use crate::browser_process::browsers::devtool_render_handler::DevToolRenderHandlerBuilder;
 #[cfg(not(target_os = "windows"))]
-use crate::browser_process::display_handler::{DisplayHandlerBuilder, SystemCursorIconSenderInner};
+use crate::browser_process::display_handler::{
+    AddressChangedSenderInner, DisplayHandlerBuilder, SystemCursorIconSenderInner,
+};
 #[cfg(not(target_os = "windows"))]
 use crate::browser_process::drag_handler::{DragHandlerBuilder, DraggableRegionSenderInner};
+#[cfg(not(target_os = "windows"))]
+use crate::browser_process::load_handler::{LoadHandlerBuilder, LoadHandlerSenderInner};
 pub use keyboard::*;
 
 pub struct WebviewBrowser {
@@ -71,6 +75,8 @@ impl Browsers {
         brp_sender: Sender<BrpMessage>,
         system_cursor_icon_sender: SystemCursorIconSenderInner,
         drag_regions_sender: DraggableRegionSenderInner,
+        load_handler_sender: LoadHandlerSenderInner,
+        address_changed_sender: AddressChangedSenderInner,
         initialize_scripts: &[String],
         _window_handle: Option<RawWindowHandle>,
     ) {
@@ -100,6 +106,8 @@ impl Browsers {
                 brp_sender,
                 system_cursor_icon_sender,
                 drag_regions_sender,
+                load_handler_sender,
+                address_changed_sender,
             )),
             Some(&uri.into()),
             Some(&BrowserSettings {
@@ -508,6 +516,8 @@ impl Browsers {
         brp_sender: Sender<BrpMessage>,
         system_cursor_icon_sender: SystemCursorIconSenderInner,
         drag_regions_sender: DraggableRegionSenderInner,
+        load_handler_sender: LoadHandlerSenderInner,
+        address_changed_sender: AddressChangedSenderInner,
     ) -> Client {
         ClientHandlerBuilder::new(RenderHandlerBuilder::build(
             webview,
@@ -516,8 +526,13 @@ impl Browsers {
             size.clone(),
             dpr,
         ))
-        .with_display_handler(DisplayHandlerBuilder::build(system_cursor_icon_sender))
+        .with_display_handler(DisplayHandlerBuilder::build(
+            webview,
+            system_cursor_icon_sender,
+            address_changed_sender,
+        ))
         .with_drag_handler(DragHandlerBuilder::build(webview, drag_regions_sender))
+        .with_load_handler(LoadHandlerBuilder::build(webview, load_handler_sender))
         .with_message_handler(JsEmitEventHandler::new(webview, ipc_event_sender))
         .with_message_handler(BrpHandler::new(brp_sender))
         .build()