+++ crates/bevy_cef_core/src/browser_process/cef_command.rs
@@ -12,8 +12,11 @@ use bevy_remote::BrpMessage;
 use raw_window_handle::RawWindowHandle;
 
 use crate::browser_process::client_handler::IpcEventRaw;
-use crate::browser_process::display_handler::SystemCursorIconSenderInner;
+use crate::browser_process::display_handler::{
+    AddressChangedSenderInner, SystemCursorIconSenderInner,
+};
 use crate::browser_process::drag_handler::DraggableRegionSenderInner;
+use crate::browser_process::load_handler::LoadHandlerSenderInner;
 use crate::browser_process::localhost::Requester;
 
 /// A `Send`-safe wrapper around [`RawWindowHandle`].
@@ -45,6 +48,8 @@ pub enum CefCommand {
         brp_sender: Sender<BrpMessage>,
         system_cursor_icon_sender: SystemCursorIconSenderInner,
         drag_regions_sender: DraggableRegionSenderInner,
+        load_handler_sender: LoadHandlerSenderInner,
+        address_changed_sender: AddressChangedSenderInner,
         initialize_scripts: Vec<String>,
         window_handle: Option<SendRawWindowHandle>,
     },
@@ -178,6 +183,8 @@ impl BrowsersProxy {
         brp_sender: Sender<BrpMessage>,
         system_cursor_icon_sender: SystemCursorIconSenderInner,
         drag_regions_sender: DraggableRegionSenderInner,
+        load_handler_sender: LoadHandlerSenderInner,
+        address_changed_sender: AddressChangedSenderInner,
         initialize_scripts: &[String],
         window_handle: Option<RawWindowHandle>,
     ) {
@@ -191,6 +198,8 @@ impl BrowsersProxy {
             brp_sender,
             system_cursor_icon_sender,
             drag_regions_sender,
+            load_handler_sender,
+            address_changed_sender,
             initialize_scripts: initialize_scripts.to_vec(),
             window_handle: window_handle.map(SendRawWindowHandle),
         });