+++ crates/bevy_cef_core/src/browser_process/cef_thread.rs
@@ -31,8 +31,11 @@ use crate::browser_process::browsers::{
 };
 use crate::browser_process::cef_command::CefCommand;
 use crate::browser_process::client_handler::{BrpHandler, IpcEventRaw, JsEmitEventHandler};
-use crate::browser_process::display_handler::{DisplayHandlerBuilder, SystemCursorIconSenderInner};
+use crate::browser_process::display_handler::{
+    AddressChangedSenderInner, DisplayHandlerBuilder, SystemCursorIconSenderInner,
+};
 use crate::browser_process::drag_handler::{DragHandlerBuilder, DraggableRegionSenderInner};
+use crate::browser_process::load_handler::{LoadHandlerBuilder, LoadHandlerSenderInner};
 use crate::browser_process::localhost::{LocalSchemaHandlerBuilder, Requester};
 use crate::browser_process::renderer_handler::{
     RenderHandlerBuilder, RenderTextureMessage, SharedDpr, SharedViewSize, TextureSender,
@@ -82,6 +85,8 @@ impl BrowsersCefSide {
                 brp_sender,
                 system_cursor_icon_sender,
                 drag_regions_sender,
+                load_handler_sender,
+                address_changed_sender,
                 initialize_scripts,
                 window_handle,
             } => {
@@ -97,6 +102,8 @@ impl BrowsersCefSide {
                     brp_sender,
                     system_cursor_icon_sender,
                     drag_regions_sender,
+                    load_handler_sender,
+                    address_changed_sender,
                     &initialize_scripts,
                     raw_handle,
                 );
@@ -165,6 +172,8 @@ impl BrowsersCefSide {
         brp_sender: Sender<BrpMessage>,
         system_cursor_icon_sender: SystemCursorIconSenderInner,
         drag_regions_sender: DraggableRegionSenderInner,
+        load_handler_sender: LoadHandlerSenderInner,
+        address_changed_sender: AddressChangedSenderInner,
         initialize_scripts: &[String],
         #[allow(deprecated)] _window_handle: Option<RawWindowHandle>,
     ) {
@@ -192,6 +201,8 @@ impl BrowsersCefSide {
                 brp_sender,
                 system_cursor_icon_sender,
                 drag_regions_sender,
+                load_handler_sender,
+                address_changed_sender,
             )),
             Some(&uri.into()),
             Some(&BrowserSettings {
@@ -503,6 +514,8 @@ impl BrowsersCefSide {
         brp_sender: Sender<BrpMessage>,
         system_cursor_icon_sender: SystemCursorIconSenderInner,
         drag_regions_sender: DraggableRegionSenderInner,
+        load_handler_sender: LoadHandlerSenderInner,
+        address_changed_sender: AddressChangedSenderInner,
     ) -> Client {
         ClientHandlerBuilder::new(RenderHandlerBuilder::build(
             webview,
@@ -510,8 +523,13 @@ impl BrowsersCefSide {
             size.clone(),
             dpr,
         ))
-        .with_display_handler(DisplayHandlerBuilder::build(system_cursor_icon_sender))
+        .with_display_handler(DisplayHandlerBuilder::build(
+            webview,
+            system_cursor_icon_sender,
+            address_changed_sender,
+        ))
         .with_drag_handler(DragHandlerBuilder::build(webview, drag_regions_sender))
+        .with_load_handler(LoadHandlerBuilder::build(webview, load_handler_sender))
         .with_message_handler(JsEmitEventHandler::new(webview, ipc_event_sender))
         .with_message_handler(BrpHandler::new(brp_sender))
         .build()