+++ src/webview.rs
@@ -253,6 +253,8 @@ fn create_webview(
     brp_sender: Res<BrpSender>,
     cursor_icon_sender: Res<SystemCursorIconSender>,
     drag_regions_sender: Res<crate::drag::DraggableRegionSender>,
+    load_handler_sender: Res<crate::navigation::LoadHandlerSender>,
+    address_changed_sender: Res<crate::navigation::AddressChangedSender>,
     webviews: Query<
         (
             Entity,
@@ -286,6 +288,8 @@ fn create_webview(
                 brp_sender.clone(),
                 cursor_icon_sender.clone(),
                 drag_regions_sender.0.clone(),
+                load_handler_sender.0.clone(),
+                address_changed_sender.0.clone(),
                 &initialize_scripts.0,
                 host_window,
             );
@@ -362,6 +366,8 @@ fn create_webview_win(
     brp_sender: Res<BrpSender>,
     cursor_icon_sender: Res<SystemCursorIconSender>,
     drag_regions_sender: Res<crate::drag::DraggableRegionSender>,
+    load_handler_sender: Res<crate::navigation::LoadHandlerSender>,
+    address_changed_sender: Res<crate::navigation::AddressChangedSender>,
     webviews: Query<
         (
             Entity,
@@ -395,6 +401,8 @@ fn create_webview_win(
                 brp_sender.clone(),
                 cursor_icon_sender.clone(),
                 drag_regions_sender.0.clone(),
+                load_handler_sender.0.clone(),
+                address_changed_sender.0.clone(),
                 &initialize_scripts.0,
                 host_window,
             );